Move styled component file
